+++ Scripts/Domain/Managers/ScoreManager.cs
@@ -19,7 +19,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 The author can be contacted at pan.i.githubcontact@gmail.com
 */
 
-using System.Diagnostics;
+using System;
 using Godot;
 
 namespace Snake.Scripts.Domain.Managers;